Moved sort_by_similarity, find_best_match and find_best_match_with_margin from util to matching
