+++ picard/matching.py
@@ -41,9 +41,13 @@ from dataclasses import (
     dataclass,
     field,
 )
+from operator import attrgetter
 from typing import (
     TYPE_CHECKING,
+    Generic,
+    Protocol,
     TypeAlias,
+    TypeVar,
 )
 
 from picard import tagger_instance
@@ -88,6 +92,12 @@ _RELEASE_WEIGHT_KEYS = {
 }
 
 
+class Similar(Protocol):
+    """Any type that provides a "similarity" attribute."""
+
+    similarity: float
+
+
 @dataclass
 class SimMatchTrack:
     """Holds the similarity score and release/track data for a matched track."""
@@ -104,6 +114,17 @@ class SimMatchRelease:
 
     similarity: float
     release: dict | None
+
+
+# Generic type variable that implements the Similar protocol.
+S = TypeVar('S', bound=Similar)
+
+
+@dataclass
+class MatchResult(Generic[S]):
+    similarity: float
+    result: S
+    reason: str | None = None
 
 
 # Type for a pair of score and weight (e.g. (0.8, 12))
@@ -356,6 +377,64 @@ def compare_to_track(metadata: 'Metadata', track: dict, weights: TieredWeights) 
             rg = release['release-group'] if "release-group" in release else None
             result = SimMatchTrack(similarity=sim, releasegroup=rg, release=release, track=track)
     return result
+
+
+def sort_by_similarity(candidates: Iterable[S]) -> list[S]:
+    """Sorts the objects in candidates by similarity.
+
+    Args:
+        candidates: Iterable with objects having a `similarity` attribute
+    Returns: List of candidates sorted by similarity (highest similarity first)
+    """
+    return sorted(candidates, reverse=True, key=attrgetter('similarity'))
+
+
+def find_best_match(candidates: Iterable[S], no_match: S) -> MatchResult[S]:
+    """Returns a MatchResult based on the similarity of candidates.
+
+    Args:
+        candidates: Iterable with objects having a `similarity` attribute
+        no_match: Match to return if there was no candidate
+
+    Returns: `MatchResult` with the similarity and the matched object as result.
+    """
+    best_match = max(candidates, key=attrgetter('similarity'), default=no_match)
+    return MatchResult(similarity=getattr(best_match, 'similarity', 0.0), result=best_match)
+
+
+def find_best_match_with_margin(
+    candidates: Iterable[S], no_match: S, min_similarity: float = 0.0, min_margin: float = 0.0
+) -> MatchResult[S]:
+    """Find best match, flagging if below floor or margin is too small.
+
+    Args:
+        candidates: Iterable with objects having a `similarity` attribute
+        no_match: Match to return if no candidate passes the floor
+        min_similarity: Reject if best score is below this floor
+        min_margin: Flag as ambiguous if best - second_best < this value
+            (skipped when there's only one candidate)
+
+    Returns: `MatchResult` with similarity, result, and reason.
+        reason is None (confident), 'ambiguous' (margin too small,
+        best match still returned), or 'below_floor' (no_match returned).
+    """
+    best = no_match
+    second_best_sim = -1.0
+    for candidate in candidates:
+        sim = candidate.similarity
+        if sim > best.similarity:
+            second_best_sim = best.similarity
+            best = candidate
+        elif sim > second_best_sim:
+            second_best_sim = sim
+
+    if best.similarity < min_similarity:
+        return MatchResult(similarity=best.similarity, result=no_match, reason='below_floor')
+
+    if second_best_sim >= 0 and (best.similarity - second_best_sim) < min_margin:
+        return MatchResult(similarity=best.similarity, result=best, reason='ambiguous')
+
+    return MatchResult(similarity=best.similarity, result=best, reason=None)
 
 
 def length_score(a: int | None, b: int | None) -> float:
